# backend/src/main/resources/openapi/investment-api.yaml
openapi: 3.0.3
info:
  title: Investment Portfolio Management API
  version: 0.1.0
  description: REST API for portfolio, watchlist, goal, news, research, and AI analysis workflows.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/v1/auth/register:
    post:
      security: []
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Access token response. Refresh session is set as an HttpOnly cookie.
          headers:
            Set-Cookie:
              description: HttpOnly refresh cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/login:
    post:
      security: []
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Access token response. Refresh session is set as an HttpOnly cookie.
          headers:
            Set-Cookie:
              description: HttpOnly refresh cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/refresh:
    post:
      security: []
      summary: Rotate a valid refresh session and issue a new access token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access token. Refresh cookie is rotated on success.
          headers:
            Set-Cookie:
              description: Rotated HttpOnly refresh cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid, expired, or revoked refresh session
...
  /api/v1/research/{symbol}/profile:
    get:
      summary: Fetch point-in-time company profile metadata for research
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Research profile snapshot with provenance, coverage, and warnings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchResponse'
  /api/v1/research/{symbol}/parameters:
    get:
      summary: Fetch source-mapped research parameters for a symbol
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Research parameter observations with evidence status and effective weights
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchResponse'
  /api/v1/research/{symbol}/score:
    get:
      summary: Fetch governed research scoring summary for a symbol
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Governed scorecard. Overall score may be null when coverage is insufficient.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchResponse'
  /api/v1/research/{symbol}/runs:
    post:
      summary: Calculate and persist a private point-in-time research snapshot
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Immutable research snapshot, owned by the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordedResearchRun'
  /api/v1/research/{symbol}/runs/latest:
    get:
      summary: Fetch the authenticated user's latest saved research snapshot
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Latest private research snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersistedResearchRun'
        '404':
          description: No saved research run exists for this symbol
...
  /api/v1/portfolios/{portfolioId}/transactions/import/preview:
    post:
      summary: Preview transaction CSV import (Premium/Admin)
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionImportRequest'
      responses:
        '200':
          description: Import preview with normalized rows and validation issues
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionImportPreview'
        '403':
          description: Premium role required
  /api/v1/portfolios/{portfolioId}/transactions/positions/{symbol}/transfer:
    patch:
      summary: Move a symbol's full transaction history to another owned portfolio
      description: Preserves cost basis by reassigning existing transaction records. Does not create synthetic BUY or SELL events.
      parameters:
        - name: portfolioId
          in: path
          required: true
          description: Source portfolio ID
          schema:
            type: string
            format: uuid
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferPositionRequest'
      responses:
        '200':
          description: Transaction history moved, audited, and both portfolios scheduled for recalculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferPositionResult'
        '400':
          description: Invalid symbol or identical source and target portfolios
        '404':
          description: Portfolio or symbol history not found for the authenticated user
  /api/v1/portfolios/{portfolioId}/transactions/import/commit:
    post:
      summary: Commit transaction CSV import (Premium/Admin)
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionImportRequest'
      responses:
        '201':
          description: Transactions created and audited
        '400':
          description: Preview validation failed
        '403':
          description: Premium role required
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
        fullName:
          type: string
          minLength: 2
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: Optional compatibility field. Preferred production flow uses the HttpOnly refresh cookie.
    AuthResponse:
      type: object
      required: [token, tokenType]
      properties:
        token:
          type: string
        tokenType:
          type: string
          example: Bearer
    TransactionImportRequest:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          maxItems: 500
          items:
            type: object
            required: [assetSymbol, type, quantity, price, currency, transactionDate]
            properties:
              assetSymbol: { type: string }
              assetName: { type: string }
              type: { type: string, enum: [BUY, SELL, MUA, BAN] }
              quantity: { type: number }
              price: { type: number }
              currency: { type: string }
              transactionDate: { type: string, format: date }
              notes: { type: string }
              fee: { type: number }
    TransferPositionRequest:
      type: object
      required: [targetPortfolioId]
      properties:
        targetPortfolioId:
          type: string
          format: uuid
    TransferPositionResult:
      type: object
      required: [sourcePortfolioId, targetPortfolioId, symbol, transactionCount]
      properties:
        sourcePortfolioId:
          type: string
          format: uuid
        targetPortfolioId:
          type: string
          format: uuid
        symbol:
          type: string
        transactionCount:
          type: integer
    TransactionImportIssue:
      type: object
      required: [row, field, message]
      properties:
        row:
          type: integer
        field:
          type: string
        message:
          type: string
    TransactionImportNormalizedRow:
      type: object
      required: [assetSymbol, assetName, type, quantity, price, currency, transactionDate, fee]
      properties:
        assetSymbol: { type: string }
        assetName: { type: string }
        type: { type: string, enum: [BUY, SELL] }
        quantity: { type: number }
        price: { type: number }
        currency: { type: string }
        transactionDate: { type: string, format: date }
        notes: { type: string, nullable: true }
        fee: { type: number }
    TransactionImportPreview:
      type: object
      required: [totalRows, validRows, invalidRows, rows, issues, readyToImport]
      properties:
        totalRows:
          type: integer
        validRows:
          type: integer
        invalidRows:
          type: integer
        rows:
          type: array
          items:
            $ref: '#/components/schemas/TransactionImportNormalizedRow'
        issues:
          type: array
          items:
            $ref: '#/components/schemas/TransactionImportIssue'
        readyToImport:
          type: boolean
...
    ResearchCompanyProfile:
      type: object
      properties:
        symbol: { type: string }
        primaryIndustry: { type: string, nullable: true }
        subIndustry: { type: string, nullable: true }
        country: { type: string, nullable: true }
        exchange: { type: string, nullable: true }
        businessModel: { type: string, nullable: true }
    ResearchObservation:
      type: object
      required: [parameterCode, moduleId, baseWeight, state, evidenceStatus, dataQuality, effectiveWeight, warnings, citation]
      properties:
        parameterCode: { type: string }
        moduleId: { type: string }
        baseWeight: { type: number }
        rawValue:
          type: number
          nullable: true
        normalizedScore:
          type: number
          nullable: true
        state:
          type: string
          enum: [ACTIVE, WATCH, DORMANT, DEPRECATED, REACTIVATED]
        evidenceStatus:
          type: string
          enum: [VERIFIED, ESTIMATED, PROXY, MISSING, CONFLICTED]
        dataQuality: { type: number }
        effectiveWeight: { type: number }
        warnings:
          type: array
          items:
            type: string
        citation:
          $ref: '#/components/schemas/ResearchEvidenceCitation'
    ResearchEvidenceCitation:
      type: object
      required: [sourceName, observedAt, sourceFields]
      properties:
        sourceName: { type: string }
        sourceUrl: { type: string, nullable: true, format: uri }
        observedAt: { type: string, format: date-time, nullable: true }
        sourceFields:
          type: array
          items: { type: string }
    ResearchStockScore:
      type: object
      properties:
        qualityScore: { type: number, nullable: true }
        growthScore: { type: number, nullable: true }
        valuationScore: { type: number, nullable: true }
        resilienceScore: { type: number, nullable: true }
        governanceScore: { type: number, nullable: true }
        catalystScore: { type: number, nullable: true }
        marketScore: { type: number, nullable: true }
        overallScore: { type: number, nullable: true }
        confidence: { type: number }
        coverage: { type: number }
        vetoStatus: { type: string }
        finalClassification: { type: string }
        coveredWeight: { type: number }
        governanceStatus: { type: string }
    ResearchResponse:
      type: object
      required: [symbol, asOf, methodologyVersion, dataVersion, profile, coverage, confidence, status, warnings, parameterResults, vetoes, stockScore]
      properties:
        symbol: { type: string }
        asOf:
          type: string
          format: date-time
        methodologyVersion: { type: string }
        dataVersion: { type: string }
        profile:
          $ref: '#/components/schemas/ResearchCompanyProfile'
        coverage: { type: number }
        confidence: { type: number }
        status: { type: string }
        warnings:
          type: array
          items:
            type: string
        parameterResults:
          type: array
          items:
            $ref: '#/components/schemas/ResearchObservation'
        vetoes:
          type: array
          items:
            type: string
        stockScore:
          $ref: '#/components/schemas/ResearchStockScore'
    RecordedResearchRun:
      type: object
      required: [runId, response]
      properties:
        runId: { type: string, format: uuid }
        response:
          $ref: '#/components/schemas/ResearchResponse'
    PersistedResearchObservation:
      type: object
      required: [parameterCode, state, evidenceStatus, dataQuality, effectiveWeight, warnings, citation]
      properties:
        parameterCode: { type: string }
        rawValue: { type: object, nullable: true, additionalProperties: true }
        normalizedScore: { type: number, nullable: true }
        state: { type: string }
        evidenceStatus: { type: string }
        dataQuality: { type: number }
        effectiveWeight: { type: number }
        warnings:
          type: array
          items: { type: string }
        citation:
          $ref: '#/components/schemas/ResearchEvidenceCitation'
    PersistedResearchRun:
      type: object
      required: [id, symbol, asOf, methodologyVersion, dataVersion, status, coverage, confidence, warnings, observations]
      properties:
        id: { type: string, format: uuid }
        symbol: { type: string }
        asOf: { type: string, format: date-time }
        methodologyVersion: { type: string }
        dataVersion: { type: string }
        status: { type: string }
        coverage: { type: number }
        confidence: { type: number }
        warnings:
          type: array
          items: { type: string }
        observations:
          type: array
          items:
            $ref: '#/components/schemas/PersistedResearchObservation'
...